+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Lumber - Mongo schema generates hasMany: array of `_id` referencing other collection.
+
 ### Changed
 - Templates - Replace unconventional `bin/www` script with `server.txt`.